+++ example/lib/gallery/gallery_page.dart
@@ -59,6 +59,8 @@ class GalleryPage extends StatelessWidget {
               const SizedBox(height: 32),
               const ScreenTransitionsSection(),
               const SizedBox(height: 32),
+              const BottomSheetsSection(),
+              const SizedBox(height: 32),
               const DialogsSection(),
               const SizedBox(height: 32),
               const IterableOptionsSection(),