+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - feat: add `NesIcons.play`
  - feat: add `NesIcons.plause`
  - feat: add `NesIcons.upload`
+ - feat: add `NesBottomSheet`
 
 # 0.16.0
  - feat: add `NesIcons.camera`